# app/rest/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Microservice for green waves management for traffic lights",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/greenwave/optimize": {
            "post": {
                "description": "Requests the optimization of green waves for traffic lights configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Optimize"
                ],
                "summary": "Request optimization",
                "parameters": [
                    {
                        "description": "Traffic lights configuration",
                        "name": "POST-body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.OptimizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.OptimizeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/codes.Error400"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/codes.Error500"
                        }
                    }
                }
            }
        },
        "/api/maxpressure/run": {
            "post": {
                "description": "Model version 2 uses explicit turning proportions and per-movement queues. Adaptive and fixed-plan metrics are evaluated separately.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaxPressure"
                ],
                "summary": "Compare movement-queue max-pressure controllers",
                "parameters": [
                    {
                        "description": "MP simulation request",
                        "name": "POST-body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MPRunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MPRunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/codes.Error400"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/codes.Error400"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "codes.Error400": {
            "type": "object",
            "properties": {
                "Error": {
                    "description": "Error indicates the error message for bad requests.",
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "codes.Error500": {
            "type": "object",
            "properties": {
                "Error": {
                    "description": "Error indicates the error message for internal server errors.",
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "dto.GreenWaveDTO": {
            "type": "object",
            "properties": {
                "band_width": {
                    "description": "Bandwidth of the green wave in seconds.",
                    "type": "number"
                },
                "distance": {
                    "description": "Distance in meters between the two junctions.",
                    "type": "number"
                },
                "interval_jun_one": {
                    "description": "Green interval on the first junction.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.GreenIntervalDTO"
                        }
                    ]
                },
                "interval_jun_two": {
                    "description": "Green interval on the second junction.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.GreenIntervalDTO"
                        }
                    ]
                },
                "travel_time": {
                    "description": "Travel time in seconds between the two junctions.",
                    "type": "number"
                }
            }
        },
        "dto.JunctionDTO": {
            "type": "object",
            "properties": {
                "cycle": {
                    "description": "Cycle is a list of phases that define the traffic light cycle for this junction",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PhaseDTO"
                    }
                },
                "id": {
                    "description": "Traffic light identifier",
                    "type": "integer"
                },
                "label": {
                    "description": "User defined alias",
                    "type": "string"
                },
                "offset": {
                    "description": "Offset of the cycle",
                    "type": "integer"
                },
                "point": {
                    "description": "Location of the junction",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.PointDTO"
                        }
                    ]
                },
                "total_duration": {
                    "description": "Total duration of the cycle in seconds",
                    "type": "integer"
                }
            }
        },
...
        "dto.ThroughGreenWaveDTO": {
            "type": "object",
            "properties": {
                "bandwidth": {
                    "description": "Bandwidth of the green wave, which is the minimum duration of the green intervals",
                    "type": "number"
                },
                "depth": {
                    "description": "Number of junctions which could be passed through",
                    "type": "integer"
                },
                "intervals": {
                    "description": "Intervals of green lights that form the green wave",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GreenIntervalDTO"
                    }
                }
            }
        },
...
        "rest.OptimizeRequest": {
            "type": "object",
            "properties": {
                "desired_speed_kmh": {
                    "description": "Desired speed in km/h for calculating green waves",
                    "type": "number"
                },
                "junctions": {
                    "description": "List of junctions with their phases and signals",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.JunctionDTO"
                    }
                },
                "optimizer_params": {
                    "description": "Contains parameters for the optimizer",
                    "type": "object",
                    "additionalProperties": true
                },
                "optimizer_type": {
                    "description": "Specifies which optimizer to use",
                    "type": "string"
                },
                "direction": {
                    "type": "string",
                    "enum": [
                        "forward",
                        "bidirectional"
                    ],
                    "description": "Calculation direction. Defaults to forward. Both directions use one program and offset per junction."
                },
                "group_ids": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    },
                    "description": "Forward signal group ID keyed by junction ID. Legacy requests without a selection use group 0."
                },
                "reverse_group_ids": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    },
                    "description": "Optional reverse signal group ID keyed by junction ID. Missing entries use the forward group. Selected groups must exist in every phase."
                }
            }
        },
        "rest.OptimizeResponse": {
            "type": "object",
            "properties": {
                "best_offsets": {
                    "description": "Contains the optimal offsets for each junction",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "green_waves": {
                    "description": "List of segments of green waves between junctions considering the optimal offsets",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/dto.GreenWaveDTO"
                        }
                    }
                },
                "optimizer_extra": {
                    "description": "Additional information about the optimization process",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.OptimizerExtra"
                        }
                    ]
                },
                "through_green_waves": {
                    "description": "List of through green waves (so they can be passed through multiple junctions) considering the optimal offsets",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ThroughGreenWaveDTO"
                    }
                },
                "reverse_green_waves": {
                    "description": "Reverse direction waves, ordered from the last junction to the first; populated for bidirectional requests.",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/dto.GreenWaveDTO"
                        }
                    }
                },
                "reverse_through_green_waves": {
                    "description": "Reverse direction waves, ordered from the last junction to the first; populated for bidirectional requests.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ThroughGreenWaveDTO"
                    }
                }
            }
        },
        "rest.OptimizerExtra": {
            "type": "object",
            "properties": {
                "fitness_history": {
                    "description": "Contains the fitness evolution over generations\nWill be represented in case of genetic algorithm\nEach value is the best fitness of the population in that generation",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "dto.MPDemandDTO": {
            "type": "object",
            "properties": {
                "base_rates": {
                    "description": "BaseRates maps link ID to base demand intensity in veh/h. Used for type \"peak\".",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "peak_duration_s": {
                    "description": "PeakDurationS is the peak phase duration in seconds. Used for type \"peak\".",
                    "type": "number"
                },
                "peak_factor": {
                    "description": "PeakFactor is the demand multiplier at peak. Defaults to 1.3 when 0. Used for type \"peak\".",
                    "type": "number"
                },
                "ramp_down_s": {
                    "description": "RampDownS is the ramp-down phase duration in seconds. Used for type \"peak\".",
                    "type": "number"
                },
                "ramp_up_s": {
                    "description": "RampUpS is the ramp-up phase duration in seconds. Used for type \"peak\".",
                    "type": "number"
                },
                "rates": {
                    "description": "Rates maps link ID to constant demand intensity in veh/h. Used for type \"constant\".",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "type": {
                    "description": "Type is the demand model: \"constant\" or \"peak\".",
                    "type": "string"
                }
            }
        },
        "dto.MPDrainDTO": {
            "type": "object",
            "properties": {
                "rates": {
                    "description": "Rates maps link ID to drain intensity in veh/h. Used for type \"rate\".",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "type": {
                    "description": "Type is the drain model: \"auto\", \"none\", or \"rate\".",
                    "type": "string"
                }
            }
        },
        "dto.MPIntersectionConfigDTO": {
            "type": "object",
            "properties": {
                "clearance_s": {
                    "type": "number"
                },
                "group_connectors": {
                    "description": "GroupConnectors maps signal group ID to the meso connector link IDs it controls.\nJSON keys are group IDs as strings (e.g. \"0\", \"1\").",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "junction": {
                    "description": "Junction is the full signal plan for this intersection.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.JunctionDTO"
                        }
                    ]
                },
                "macro_node_id": {
                    "description": "MacroNodeID is the macro-graph node ID identifying this intersection.",
                    "type": "integer"
                },
                "max_green_s": {
                    "type": "number"
                },
                "min_green_s": {
                    "description": "Timing is an optional practical extension. Omitted values use 5/60/5 seconds.\nExplicit zeros disable the respective bound, for the paper's ideal controller.",
                    "type": "number"
                }
            }
        },
...
        "dto.MPNetworkDTO": {
            "type": "object",
            "properties": {
                "links": {
                    "description": "Links is the list of meso links (road segments and connectors).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MPLinkDTO"
                    }
                }
            }
        },
        "dto.MPProposalDTO": {
            "type": "object",
            "properties": {
                "junction": {
                    "$ref": "#/definitions/dto.JunctionDTO"
                },
                "macro_node_id": {
                    "type": "integer"
                }
            }
        },
        "dto.MPRunRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "Config holds simulation parameters.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.MPSimConfigDTO"
                        }
                    ]
                },
                "coordinated_corridors": {
                    "description": "CoordinatedCorridors contains directed road-link paths, each with at least\nthree links and assigned signal-controlled movements between them.\nRequired when config.alpha > 0; reverse direction needs its own path.",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "demand": {
                    "description": "Demand defines the traffic input per entry link.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.MPDemandDTO"
                        }
                    ]
                },
                "drain": {
                    "description": "Drain controls how vehicles leave the network at boundary exits.\nOmit or set type to \"auto\" for the default behaviour (zero queue each step).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.MPDrainDTO"
                        }
                    ]
                },
                "initial_backlogs": {
                    "description": "InitialBacklogs supplies queues outside entry roads.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "initial_movement_queues": {
                    "description": "InitialMovementQueues supplies separate queues by connector ID, instead of road totals.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "initial_queues": {
                    "description": "InitialQueues sets starting queue length (vehicles) per link ID. Optional.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "intersections": {
                    "description": "Intersections defines the signal configuration and group-to-connector mapping.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MPIntersectionConfigDTO"
                    }
                },
                "network": {
                    "description": "Network is the meso-level graph (road segments + connector links).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.MPNetworkDTO"
                        }
                    ]
                },
                "turning_ratios": {
                    "description": "TurningRatios maps connector IDs to arrival proportions; every branching road must sum to 1.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                }
            }
        },
        "dto.MPRunResponse": {
            "type": "object",
            "properties": {
                "baseline": {
                    "$ref": "#/definitions/maxpressure.Evaluation"
                },
                "evaluation": {
                    "$ref": "#/definitions/maxpressure.Evaluation"
                },
                "model_version": {
                    "type": "integer"
                },
                "proposal": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MPProposalDTO"
                    }
                },
                "proposal_accepted": {
                    "type": "boolean"
                },
                "proposal_evaluation": {
                    "$ref": "#/definitions/maxpressure.Evaluation"
                },
                "standard_mp": {
                    "$ref": "#/definitions/maxpressure.Evaluation"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.MPSimConfigDTO": {
            "type": "object",
            "properties": {
                "alpha": {
                    "description": "Alpha lies in [0,1]; the coordination weight is Alpha * Q^2, Q in veh/step.",
                    "type": "number"
                },
                "delta_t": {
                    "description": "DeltaT is the simulation step duration in seconds. Defaults to 1.0 if not set.",
                    "type": "number"
                },
                "sim_time": {
                    "description": "SimTime is the total simulation duration in seconds. Must be > 0.",
                    "type": "number"
                },
                "storage_model": {
                    "description": "StorageModel is \"point_queue\" (default, Xu) or \"finite_storage\" (extension).",
                    "type": "string"
                }
            }
        },
...
    }
}